PET-105 Refactor: MemberController로 통일
